alloc: Convert `not(no_global_oom_handling)` to cargo feature.
To better support public use of this feature, we convert it to a Cargo
"global-oom-handling", on by default. Note that the meaning is inverted,
and there is no more need to negation. Besides simplifying things, this
fixes the monotonicity of the feature, i.e. the property that turning on
more features is always safe.